requirements.txt
# NexaCV — Python Dependencies
flask>=3.0.0
werkzeug>=3.0.0
groq>=0.9.0
pdfplumber>=0.11.0
spacy>=3.7.0
reportlab>=4.2.0
bcrypt>=4.1.3
python-dotenv>=1.0.1
# After install run: python -m spacy download en_core_web_sm
